+++ packages/spacecat-shared-drs-client/test/setup-env.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Adobe. All rights reserved.
+ * Copyright 2026 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -9,8 +9,8 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-
-export type * from './errors';
-export type * from './models';
-export type * from './util';
-export type { DataAccess } from './service';
+// eslint-disable-next-line no-console
+console.log('Forcing HTTP/1.1 for Adobe Fetch');
+process.env.HELIX_FETCH_FORCE_HTTP1 = 'true';
+process.env.AWS_XRAY_SDK_ENABLED = 'false';
+process.env.AWS_XRAY_CONTEXT_MISSING = 'IGNORE_ERROR';