+++ packages/spacecat-shared-vault-secrets/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [@adobe/spacecat-shared-vault-secrets-v1.1.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-vault-secrets-v1.0.1...@adobe/spacecat-shared-vault-secrets-v1.1.0) (2026-03-04)
+
+### Features
+
+* add dev alias secret overrides to spacecat-shared-vault-secrets ([#1400](https://github.com/adobe/spacecat-shared/issues/1400)) ([93b7c67](https://github.com/adobe/spacecat-shared/commit/93b7c670d301d8ac7eae12e2e903e50541fd9f44))
+
 ## [@adobe/spacecat-shared-vault-secrets-v1.0.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-vault-secrets-v1.0.0...@adobe/spacecat-shared-vault-secrets-v1.0.1) (2026-03-02)
 
 ### Bug Fixes